+++ hc/api/transports.py
@@ -877,10 +877,21 @@ class Telegram(HttpTransport):
 class Sms(HttpTransport):
     URL = "https://api.twilio.com/2010-04-01/Accounts/%s/Messages.json"
 
+    class ErrorModel(BaseModel):
+        code: int
+
     @classmethod
     def raise_for_response(cls, response: curl.Response) -> NoReturn:
         if response.status_code == 400:
+            try:
+                doc = Sms.ErrorModel.model_validate_json(response.content, strict=True)
+                if doc.code == 21211:
+                    raise TransportError("Invalid phone number", permanent=True)
+            except ValidationError:
+                pass
+
             logger.debug("Twilio Messages HTTP 400 with body: %s", response.content)
+
         raise TransportError(f"Received status code {response.status_code}")
 
     def is_noop(self, check: Check) -> bool:
@@ -920,9 +931,19 @@ class Sms(HttpTransport):
 class Call(HttpTransport):
     URL = "https://api.twilio.com/2010-04-01/Accounts/%s/Calls.json"
 
+    class ErrorModel(BaseModel):
+        code: int
+
     @classmethod
     def raise_for_response(cls, response: curl.Response) -> NoReturn:
         if response.status_code == 400:
+            try:
+                doc = Call.ErrorModel.model_validate_json(response.content, strict=True)
+                if doc.code == 21211:
+                    raise TransportError("Invalid phone number", permanent=True)
+            except ValidationError:
+                pass
+
             logger.debug("Twilio Calls HTTP 400 with body: %s", response.content)
         raise TransportError(f"Received status code {response.status_code}")
 
@@ -959,10 +980,23 @@ class Call(HttpTransport):
 class WhatsApp(HttpTransport):
     URL = "https://api.twilio.com/2010-04-01/Accounts/%s/Messages.json"
 
+    class ErrorModel(BaseModel):
+        code: int
+
     @classmethod
     def raise_for_response(cls, response: curl.Response) -> NoReturn:
         if response.status_code == 400:
+            try:
+                doc = WhatsApp.ErrorModel.model_validate_json(
+                    response.content, strict=True
+                )
+                if doc.code == 21211:
+                    raise TransportError("Invalid phone number", permanent=True)
+            except ValidationError:
+                pass
+
             logger.debug("WhatsApp HTTP 400 with body: %s", response.content)
+
         raise TransportError(f"Received status code {response.status_code}")
 
     def is_noop(self, check: Check) -> bool: